pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "readwater"
version = "0.1.0"
description = "AI-powered inshore saltwater fishing trip planner using multi-scale satellite imagery analysis"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/readwater"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = ["integration: requires API keys and network access"]

[tool.ruff]
target-version = "py311"
line-length = 100
